requirements.txt
celery==4.3.0
Django==2.1.11
psycopg2-binary==2.7.6.1
whitenoise==4.1.2
djangorestframework==3.8.2
#django-rest-swagger==2.1.2
#drf-openapi==1.0.0
drf-yasg[validation]==1.9.1
#waitress==1.0.2
django-oauth-toolkit==1.0.0
django-cors-middleware==1.3.1
django-elasticsearch-dsl==0.2.0
elasticsearch-dsl==5.3.0
dj-database-url==0.4.2
social-auth-core==1.5.0
social-auth-app-django==2.0.0
django-extensions==1.9.0
channels==2.1.2
#daphne==2.2.2
gunicorn==19.9.0
uvicorn==0.8.4
asgi_redis==1.4.3
drf-writable-nested==0.5.1
drf-extensions==0.3.1
termcolor==1.1.0
faker==0.8.10
tqdm==4.19.5
django-debug-toolbar==1.9.1
asgiref==2.3.2
selenium==3.141.0
Pillow==5.0.0
django-storages[azure,google]==1.7.1
boto3==1.9.86
django-querycount==0.7.0
urllib3<1.25,>=1.21.1
bpython==0.18
factory_boy==2.12.0
watchdog==0.9.0
